requirements.txt
click
rich
pyOpenSSL
semgrep
rsa
